Refactoring to accommodate multiple output formats
