+++ src/templates/typescript/complexMemberTemplate.ts
