+++ src/templates/typescript/primitiveMemberTemplate.ts
