+++ src/templates/typescript/registerPartials.ts
